+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 实现保护机制，并开发预设保护器。
   - com.dwarfeng.acckeeper.impl.handler.protector.AbstractProtectorRegistry。
   - com.dwarfeng.acckeeper.impl.handler.protector.DoNothingProtectorRegistry。
+  - com.dwarfeng.acckeeper.impl.handler.protector.PasswordProtectorRegistry。
 
 - 实现运维指令。
   - com.dwarfeng.acckeeper.impl.service.telqos.LoginCommand。